adds test for InvocationTargetException deserialization
